+++ src/Menu/HouseChoiceInfoMenu.cpp
@@ -41,7 +41,7 @@ HouseChoiceInfoMenu::HouseChoiceInfoMenu(HOUSETYPE newHouse) : MentatMenu(HOUSET
         case HOUSETYPE::HOUSE_SARDAUKAR:    anim = pGFXManager->getAnimation(Anim_SardaukarPlanet); break;
         case HOUSETYPE::HOUSE_MERCENARY:    anim = pGFXManager->getAnimation(Anim_MercenaryPlanet); break;
         default: {
-            THROW(std::invalid_argument, "HouseChoiceInfoMenu::HouseChoiceInfoMenu(): Invalid house id '%d'.", newHouse);
+            THROW(std::invalid_argument, "HouseChoiceInfoMenu::HouseChoiceInfoMenu(): Invalid house id '%d'.", static_cast<int>(newHouse));
         } break;
     }
     // clang-format on
